DIV thresh, max and min are floats, vectorise fake_B_discrete computation

+++ data/geo_unpickler.py
@@ -155,7 +155,7 @@ class GeoUnpickler(object):
 			if not key in data_dict.keys():
 				continue
 
-			data_dict[key] = torch.DoubleTensor([data_dict[key]])
+			data_dict[key] = torch.FloatTensor([data_dict[key]])
 
 		for key in ['conn_comp_hist']:
 			if not key in data_dict.keys():
